% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fun.R
\name{cancel_all_orders}
\alias{cancel_all_orders}
\title{Cancel all open orders}
\usage{
cancel_all_orders()
}
\description{
Cancel a particular open order (or set of open orders) by \code{txid}.
A list of open order can be obtained with \code{get_open_orders()}.
}
\examples{
cancel_all_orders()
}
