% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fun.R
\name{cancel_all_orders_afterX}
\alias{cancel_all_orders_afterX}
\title{Cancel all open orders after X}
\usage{
cancel_all_orders_afterX(timeout)
}
\arguments{
\item{timeout}{Timeout in seconds}
}
\description{
This function provides a "Dead Man's Switch" mechanism to protect the client from network malfunction, extreme latency or unexpected matching engine downtime. The client can send a request with a timeout (in seconds), that will start a countdown timer which will cancel all client orders when the timer expires. The client has to keep sending new requests to push back the trigger time, or deactivate the mechanism by specifying a timeout of 0. If the timer expires, all orders are cancelled and then the timer remains disabled until the client provides a new (non-zero) timeout.
}
