% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_fun.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel open order}
\usage{
cancel_order(txid)
}
\arguments{
\item{txid}{Open order transaction ID (txid). A list of open order can be obtained with \code{get_open_orders()}.}
}
\value{
Confirmation message (string)
}
\description{
Cancel a particular open order (or set of open orders) by \code{txid}.
}
