% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_all_ohlc}
\alias{get_all_ohlc}
\title{Get all OHLC data}
\usage{
get_all_ohlc(since, interval, qfilter = NULL, timeout = 2)
}
\arguments{
\item{since}{Return up to 720 OHLC data points since given timestamp.}

\item{interval}{Time intervals in minutes.}

\item{qfilter}{Filter to select a quote pair (Options: \code{ZEUR}, \code{ZUSD}, \code{ZAUD}, \code{XETH}, \code{XXBT}, \code{USDT}, \code{ZJPY}, \code{CHF}, \code{DAI}, \code{USDC}, \code{ZCAD}, \code{DOT}).}

\item{timeout}{Timeout in seconds between API requests.}
}
\value{
OHLC data (dataframe)
}
\description{
Get OHLC data for all available asset pairs
Note: the last entry in the OHLC array is for the current, not-yet-committed frame and will always be present, regardless of the value of \code{since}.
}
\examples{
get_ohlc("XBTUSD", "2022-01-01", 1440)
}
