% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_asks}
\alias{get_asks}
\title{Get asks}
\usage{
get_asks(pair)
}
\arguments{
\item{pair}{Asset pair \code{id} or \code{altname}.}
}
\value{
Asks (dataframe)
}
\description{
Returns dataframe of asks in order book for an asset pair.
}
\examples{
get_bids("XBTUSD")
}
