% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_available_balance}
\alias{get_available_balance}
\title{Get balance}
\usage{
get_available_balance()
}
\value{
Balance (dataframe)
}
\description{
Retrieve a summary of collateral balances, margin position valuations, equity and margin level.
}
