% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_bids}
\alias{get_bids}
\title{Get bids}
\usage{
get_bids(pair)
}
\arguments{
\item{pair}{Asset pair \code{id} or \code{altname}.}
}
\value{
Bids (dataframe)
}
\description{
Returns dataframe of bids in order book for an asset pair.
}
\examples{
get_bids("XBTUSD")
}
