% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_closed_orders}
\alias{get_closed_orders}
\title{Get closed orders}
\usage{
get_closed_orders()
}
\value{
Closed orders (dataframe)
}
\description{
Retrieve information about all closed orders (fullfilled and cancelled).
}
\examples{
get_closed_orders()
}
