% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_open_orders}
\alias{get_open_orders}
\title{Get open orders}
\usage{
get_open_orders()
}
\value{
Open orders (dataframe)
}
\description{
Retrieve information about currently open orders.
}
\examples{
get_open_orders()
}
