% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_order_info}
\alias{get_order_info}
\title{Query Orders Info}
\usage{
get_order_info(txid)
}
\value{
Order Info (dataframe)
}
\description{
Retrieve information about an order (fullfilled and cancelled).
}
