% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_spread}
\alias{get_spread}
\title{Get spread}
\usage{
get_spread(pair)
}
\arguments{
\item{pair}{Asset pair \code{id} or \code{altname}.}
}
\value{
Recent spread (dataframe)
}
\description{
Get recent spread.
}
\examples{
get_spread("XBTUSD")
}
