% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_sysstatus}
\alias{get_sysstatus}
\title{Get Kraken system status}
\usage{
get_sysstatus()
}
\value{
System status (string)
}
\description{
Get the current system status or trading mode.
}
\examples{
get_sysstatus()
}
