% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_trade_balance}
\alias{get_trade_balance}
\title{Get trade balance}
\usage{
get_trade_balance(asset)
}
\arguments{
\item{asset}{Asset.}
}
\value{
Balance (dataframe)
}
\description{
Retrieve a summary of collateral balances, margin position valuations, equity and margin level.
}
