% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_trade_history}
\alias{get_trade_history}
\title{Get trade history}
\usage{
get_trade_history(type = "all", trades = FALSE, start = NULL, end = NULL)
}
\arguments{
\item{type}{Type of trade (default = "all). Options: \code{all}, \verb{any position}, \verb{closed position}, \verb{closing position}, \verb{no position}.}

\item{trades}{Whether or not to include trades related to position in output (default = FALSE).}

\item{start}{Optional: Starting unix timestamp or trade txid of results.}

\item{end}{Optional: Ending unix timestamp or trade txid of results.}
}
\value{
Trade history (dataframe)
}
\description{
Retrieve information about trades/fills. 50 results are returned at a time, the most recent by default.
}
