% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_trades}
\alias{get_trades}
\title{Get recent trades}
\usage{
get_trades(pair)
}
\arguments{
\item{pair}{Asset pair \code{id} or \code{altname}.}
}
\value{
Recent trades (dataframe)
}
\description{
Returns the last 1000 trades.
}
\examples{
get_trades("XBTUSD")
}
