% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds_fun.R
\name{withdraw_funds}
\alias{withdraw_funds}
\title{Withdraw funds}
\usage{
withdraw_funds(asset, key, amount)
}
\arguments{
\item{asset}{Asset being withdrawn.}

\item{key}{Withdrawal key name, as set up on your account.}

\item{amount}{Amount to be withdrawn.}
}
\value{
Confirmation message (string)
}
\description{
Withdraw funds to a reference address.
}
