% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds_fun.R
\name{withdraw_status}
\alias{withdraw_status}
\title{Withdraw status}
\usage{
withdraw_status(asset)
}
\arguments{
\item{asset}{Asset being withdrawn.}
}
\value{
Confirmation message (string)
}
\description{
Retrieve information about recently requests withdrawals.
}
